requirements.txt
streamlit==1.39.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
plotly>=5.22.0
sentence-transformers>=3.0.0
torch>=2.2.0
PyPDF2>=3.0.0
python-docx>=1.1.2
spacy>=3.7.0
preshed>=3.0.9
srsly>=2.4.8
catalogue>=2.0.10
thinc>=8.2.2